ref: split the different domain lookup methods into files
