+++ .env.sample
@@ -51,6 +51,9 @@ DL_DEMO_PASS='password123'
 # DL_DOMAIN_INFO_API=''
 # DL_DOMAIN_SUBS_API=''
 
+# (optional) Use your own who-dat lookup instance (see: https://github.com/lissy93/who-dat)
+# DL_WHO_DAT_URL='https://who-dat.as93.net'
+
 # (optional) To disable write-methods being called from the UI (e.g. if no auth)
 # DL_DISABLE_WRITE_METHODS='false'
 #
